Setup materialize-css
The library's javascript cannot be imported as a CommonJS module, which
is needed by Brunch.

There doesn't seem to be an easy way to tell Brunch to include that
single file without wrapping it in a module, so a Plug was added to
serve all materialize assets from the corresponding dist directory in
node_modules.

The alternative would have been to add the files in the vendor
directory, but for the moment we prefer to have the dependency managed
by NPM.